make cache work properly with gRPC
