+++ pubspec.yaml
@@ -56,6 +56,10 @@ dependency_overrides:
     git:
       url: https://github.com/jakobkoerber/xml2json.git
       ref: master
+  grpc:
+    git:
+      url: https://github.com/jakobkoerber/grpc-dart.git
+      ref: master
 
 dev_dependencies:
   flutter_test:
